pyproject.toml
[tool.poetry]
name = "disease-trend-system"
version = "0.1.0"
description = ""
authors = ["DieNice <DailyPlanning@bk.ru>"]
license = "GNU AGPLv3"
readme = "README.md"
packages = [{include = "disease_trend_system"}]

[tool.poetry.dependencies]
python = "^3.10"
dash = "^2.9.2"
dash-extensions = "^0.1.13"
dash-bootstrap-components = "^1.4.1"
pandas = "^1.5.3"
sqlalchemy = "^2.0.8"
sqlalchemy-utils = "^0.40.0"
flask-restful = "^0.3.9"
pymysql = "^1.0.3"
cryptography = "^40.0.1"
python-dotenv = "^1.0.0"
flask-admin = "^1.6.1"
flask-login = "^0.6.2"
gunicorn = "^20.1.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
pylint = "^2.17.2"
mypy = "^1.1.1"
ipython = "^8.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
